Refactor Document as TypeScript, dropping wrapScalars option
