+++ src/tags/core.js
@@ -19,8 +19,7 @@ function intStringify(node, radix, prefix) {
 
 export const nullObj = {
   identify: value => value == null,
-  createNode: (schema, value, ctx) =>
-    ctx.wrapScalars ? new Scalar(null) : null,
+  createNode: () => new Scalar(null),
   default: true,
   tag: 'tag:yaml.org,2002:null',
   test: /^(?:~|[Nn]ull|NULL)?$/,
