+++ src/tags/json.js
@@ -22,8 +22,7 @@ export const json = [
   },
   {
     identify: value => value == null,
-    createNode: (schema, value, ctx) =>
-      ctx.wrapScalars ? new Scalar(null) : null,
+    createNode: () => new Scalar(null),
     default: true,
     tag: 'tag:yaml.org,2002:null',
     test: /^null$/,
